Rename backup -> fileBackup
